update login page
refactor
